Fix JERSEY-2123: request was rejected because no multipart boundary
After adding proxy support for agentClient and kmAuthClient,
we have the following error when sending a message:

the request was rejected because no multipart boundary was found

This is due to JERSEY-2123.

The workaround is when sending multipart data to change the second
parameter of post call with a call to MultiPartMediaTypes.createFormData()

This should create a new MediaType instance with a configured boundary,
which will be picked up by the MultiPart writer.