Create Sleeper class and add javadoc for unsupported methods.
Using Sleeper for sleeping in DatafeedEventServiceV2 class to
avoid using Powermock in testing.

Move Sleeper to services package to be package-protected.

The APIs for listing datafeeds, delete a datafeed, and get ackId
from previous read are not supported in DatafeedV1 so it should
be notified in javadoc of this APIs.