Backporting APP-2781: DatafeedV2 Implementation (#150) (#155)
* APP-2781: DatafeedV2 Implementation (#150)

* DatafeedV2 implementation

* Javadoc for DatafeedClient and DatafeedEventService

* Using ackId when reading datafeed

* Set DatafeedClient to final class

* Public factory class for instantiating DatafeedClient and DatafeedEventService

* Add unittest for DatafeedClient

* Threading strategy for DatafeedEventService

* Avoid spawning a new thread each time catching an exception

* Shutdown the threadpool properly

* Handle stale feed case

* Adding unittest for DataEventsService

* Using CountDownLatch instead of while loop in unittest for stale case

* Create Sleeper class and add javadoc for unsupported methods.

Using Sleeper for sleeping in DatafeedEventServiceV2 class to
avoid using Powermock in testing.

Move Sleeper to services package to be package-protected.

The APIs for listing datafeeds, delete a datafeed, and get ackId
from previous read are not supported in DatafeedV1 so it should
be notified in javadoc of this APIs.

(cherry picked from commit d318cf5900b8cad3491aed7d761166da2b53e669)

* Move DatafeedClient out of datafeeds package (#152)

Before, the DatafeedClient is under symphony.api package, so moving it
to datafeeds package might break the compatibility of project. That's
why it should be moved out of datafeed package, back to symphony.api.