APP-3072: Load parent URL configuration with Spring Boot
When Spring Boot was used the overriding mechanism we introduced where
you can define the host, scheme, port and context at the top level and
also override it for each sub component (agent, pod, ...) was not used.

The logic has been builtin directly in the POJOs so it is always applied
either when we use Jackson to deserialize or the Spring Boot
configuration framework.