APP-3106: Publish BOM when releasing
The nexus plugin we were using was not supporting BOM files. We ended up
using the maven-publish plugin (and kept the nexus staging plugin to
close and release the staging repository).

By default the repository URL is the Nexus staging one.

The POM definition is shared at top level with a closure.

Also applied this to legacy project.