APP-3069: Ability to directly set privateKey and certificate content (#271)
* Add internal dependencies version to bom module

* Update missing dependencies

* Adding plugin to http and template submodule

* Ability to directly set privateKey and certificate content

User can set the privateKey and certificate content in bytes array to BdkConfig
for bot and app authentication.

Only one of path or content should be configured. If user configure both of them,
an AuthInitializationException will be thrown. The exception should be thrown at the step of
initializing the Authenticator instead of loading config because the BdkConfig can be set programmatically.

Unittest added

* Fix typo and add documentation

* Update documentation