APP-3106: Publish BOM when releasing (#269)
* APP-3106: Publish BOM when releasing

The nexus plugin we were using was not supporting BOM files. We ended up
using the maven-publish plugin (and kept the nexus staging plugin to
close and release the staging repository).

By default the repository URL is the Nexus staging one.

The POM definition is shared at top level with a closure.

Also applied this to legacy project.

* Description was missing in POM files

Maven publications is evaluated too early, before description is set so
it was missing. Found this approach that works but is a bit ugly...

* Removed projectVersion parameter in gradle and jenkins files (#270)

Co-authored-by: Elias Croze <66254040+symphony-elias@users.noreply.github.com>
Co-authored-by: Thibault Pensec <39826516+symphony-thibault@users.noreply.github.com>