Generated POM for legacy bot SDK is missing versions (#310)
* Generated POM for legacy bot SDK is missing versions

Versions for Spring Boot Actuator and Micrometer were not defined in the
constraints. While it was not breaking the build because transitive
dependencies were used, it was breaking the POM generation, leaving the
versions empty.

A project using the SDK as a dependency would then not get the
transitive dependencies automatically.

* SDK APIs should be transitive dependencies

As classes from the client module or the sms sdk are public in the bot
sdk APIs we need to expose them as 'api' for them to be transitive.