APP-3328: Replace usage of deprecated exchange() in ApiClientWebClient
The method exchange() in Spring WebClient is deprecated, I replaced its
usage with exchangeToMono().

Also avoid creating a mock server client as we already have created one.
Also stopAsync the mock server for faster tests (was waiting 2s doing
nothing), we might not release the port at the end of each test but
since it is recreated each time (and now on a random port) it should be
ok.

Simplified replaceAll with replace to avoid use of regex.

And fixed many IDEA warnings related to the Junit 5 test (no need for
public class, put asserts in the proper order)