PLAT-10295: Prepare 2.0.0 release
Bump version to 2.0.0-SNAPSHOT (will be 2.0.0-RC1 in the RC branch)
Update documentation to use 2.0.0 as we expect most of of the readers to
use this version forward.

Use a branch version for Swagger spec for stability and reproductible
builds.

Upgrade dependencies (Spring Boot to 2.4.1 and others)
It caused one of the test to fail because the content type header is a
bit different, relaxed the test to pass.