Update the built-in Help command
Create a dedicated HelpCommand class for the help command take the ActivityRegistry and the MessageService as the params of constructors.

HelpCommand instance is created and registered to ActivityRegistry at the SymphonyBdk level.

Reuse name and description fields in ActivityInfo then the CommandActivityInfo.java can be removed.

Using template for help command reply.

Update Slash annotation in SpringBoot starter.