Update the built-in Help command
When the bot dev register a new command, check the command activity name, if there's any command with the same name already existing, unsubscribe it from the datafeed loop then remove it from the activity list of ActivityRegistry.

Revert the ActivityInfo class to prevent the breaking changes.

Building String programmatically for the help command message instead of using a template engine.

Update unittest.