PLAT-10360: Remove Spring warning during startup
When Spring integration is used we could see messages such as:

2021-01-08 11:40:18.767  INFO 86263 --- [           main] trationDelegate$BeanPostProcessorChecker : Bean 'com.symphony.bdk.spring.config.BdkActivityConfig' of type [com.symphony.bdk.spring.config.BdkActivityConfig] is not eligible for getting processed by all BeanPostProcessors (for example: not eligible for auto-proxying)

during startup.

This was caused because the BeanPostProcessor for slash commands has a
dependency on ActivityRegistry. This is a bean that has other beans as
dependencies too. So the processor is created after a few beans has
already been created and Spring warns us that those beans might not have
been processed.

We remove this dependency by not injecting the activity registry during
the processor's construction but by retrieving it only when needed (that
is when we start processing slash commands).

I relied on the existing tests and manual testing to check that the
warning is no longer printed.