PLAT-10379: CircleCI configuration for Snyk
We want to run Snyk only master branch to avoid exposing the SNYK_TOKEN
on any PR (as the project is OSS).
A separate job is defined for that purpose using the Gradle Snyk plugin.

The SNYK_TOKEN is set in a context that can be reused across DevX jobs.

Also configure an exclusion for log4j that we won't fix.