PLAT-10379: Run Snyk with Gradle + cleaner build
The Gradle Snyk plugin is used, it adds 2 tasks:
- snyk-test
- snyk-monitor

The task is only configured at the root level and with the option to
look at subproject (it is faster than running it for each project).

Most of the changes for the build are to use convention plugin which is
a better approach to configure builds.

In buildSrc/ we define our own plugins with the common configuration and
we can better compose them by applying then where needed.

For instance we use the publish plugin only on the projects that are
published. The root build is much simpler to read and we avoid the magic
and filtering of subprojects{}

This has not been done for the legacy project as it is ... legacy ...