PLAT-10480 : fix spring starter's extended healthCheck (#436)
This is PR is related to [PLAT-10480](https://perzoinc.atlassian.net/browse/PLAT-10480).
The extended health check was not working in Spring starter. The issue comes from the api clients configuration that injects a pod api client instead of agent's one to SystemApi that provides healthCheck. 
The bdk core is not impacted as its configuration is already correct.