Fix ConcurrentModificationException in Datafeed loop #475 (#476)
A ConcurrentModificationException can happen in the Datafeed (DF) loop upon startup.
The access to the listeners collection in the DF loop is not protected
and accessed by multiple threads.

Because of the for loop we need to synchronize on the list itself.

(cherry picked from commit c9fc39c89ac1eef816272c1e03957929efbc5083)