Support DFv2 tag parameter (#488)
The tag parameter enable running multiple instances of a bot. Each
instance will (randomly) receive events to process.

If the processing of events fail, ack id of the DF loop won't be updated
and the events will stay in the DF queue to be redispatched.

The branch is using a custom branch for the specs for now. To be changed
before merging!