Exception handling in DFv2 loop
A new exception EventException (runtime to not break compatibility) has
been introduced. It can be used by bot's developer to explicitly
re-queue events (i.e not update the ack id in the DFv2 loop).
This works only for DFv2.

This is supported in event listeners as well as in activities.

Also log the case when an event is unknown as info, this might happen in
the future if there are new Symphony events to handle.

Warn if event processing is slow

The warning is printed if event processing is ok but slow, to prevent
developers about events being re-queued.

If the processing fails it might be because of a timeout so we don't
want to clutter the logs.

Cleanup comment about tag

The tags are per user so it is very unlikely that truncating the service
account name is going to be a problem as one bot deployment = one
service account most of the time.

Document datafeed loop exceptions/instances
Document slow event processing

Better coverage for exception