Clarify stickiness on Agent LB is only needed for DFv1
DFv2 is using long polling and the agent is not storing any feedId (as for DFv1) so there’s no need to always call the same instance of the Agent for a given feedId. It is possible to read a datafeed using any Agent instance behind the LB. That's why the stickiness of a load balancer is only needed in case of DFv1.