Run multiple instances of a bot (#492)
* Support DFv2 tag parameter (#488)

The tag parameter enable running multiple instances of a bot. Each
instance will (randomly) receive events to process.

If the processing of events fail, ack id of the DF loop won't be updated
and the events will stay in the DF queue to be redispatched.

The branch is using a custom branch for the specs for now. To be changed
before merging!

* Add examples of multiple instances bot with DFv2

This is a sort of a benchmark to make sure that when we run multiple
instances of a bot, no messages are lost.

2 bots, one to send messages and check they are being replied too, one
  to receive messages and reply to them, the second one is meant to
  running multiple instances.

* Improve multiple instances example with Hazelcast

Hazelcast is used to provide a distributed cache and ensure an event
does not get processed twice.

* Remove kafka example

Replaced with DFv2 tag parameter + Hazelcast

* Remove unused uploadArchives build configuration

This is preventing to build with Gradle 7.
We don't use this flag anymore since we apply the maven publish with a
convention plugin.

* Exception handling in DFv2 loop

A new exception EventException (runtime to not break compatibility) has
been introduced. It can be used by bot's developer to explicitly
re-queue events (i.e not update the ack id in the DFv2 loop).
This works only for DFv2.

This is supported in event listeners as well as in activities.

Also log the case when an event is unknown as info, this might happen in
the future if there are new Symphony events to handle.

Warn if event processing is slow

The warning is printed if event processing is ok but slow, to prevent
developers about events being re-queued.

If the processing fails it might be because of a timeout so we don't
want to clutter the logs.

Cleanup comment about tag

The tags are per user so it is very unlikely that truncating the service
account name is going to be a problem as one bot deployment = one
service account most of the time.

Document datafeed loop exceptions/instances
Document slow event processing

Better coverage for exception