#515 Unable to create SlashHandler with same value, different @mention
Previously, we were not able to register two slash commands having the same name. This was done in order to avoid registering duplicates. However, two slash commands having the same name are not necessarily duplicates as they can differ in bot's mention required property : In fact, we can have one command requiring bot's mention while the second one is not requiring it. In this case, both of them should be registered.

The issue was coming from the fact that we only relied on command's name to check if it is not already present in the registry. The fix was to check the bot's mention required property as well.