Jersey HTTP client SSL/TLS broken configuration
When a proxy is used, the Jersey client was switch to the Apache HTTP
connection pool which was not properly configured to use custom
key/trust stores.

The connector has to be set all the time to make the Apache pool is
used. Then the SSL context must be set directly in the pool as well.

Added a unit test to check this using Mock Server.

Closes #519