Encode stream/msg id to url-safe base64 if needed
The service methods of the BDK were accepting only url-safe base64
stream or message ids. Now this escaping is performed automatically if
not already done.

It concerns mainly the stream and message services. Having unit tests
for every methods felt a bit cumbersome so I added one for each service
only.

The escapeString method in ApiClient was a good candidate for such a
change but it takes all sorts of ids that are parts of API paths so it
felt risky to change it so globally (and would probably not work for app
ids for instance that are not base64).

I considered the StreamUtil class to a public API and probably already
used by BDK users so I did not changed the existing method (i.e it
should not have been named with stream id since it can also be used for
message).

PLAT-11214