Avoid Jersey warning about unsent headers (#568)
When a request was made with Jersey and a multipart body the following
warning was printed:

WARNING: There are some request headers that have not been sent by connector [org.glassfish.jersey.apache.connector.ApacheConnector]. Probably you added those headers in WriterInterceptor or MessageBodyWriter. That feature is not supported by the connector. Please, do not modify headers in WriterInterceptor or MessageBodyWriter or use default HttpUrlConnector instead.
Unsent header changes: [MIME-Version]

This is explained here:
https://eclipse-ee4j.github.io/jersey.github.io/documentation/latest/client.html#connectors.warning

This can be avoided by setting the header explicitly in the body is
multipart to avoid it being set in the MessageBodyWriter.

As this was just printing a warning, automated tests for that is tricky.
It has been manually tested by sending messages with and without
attachments.