Prevent invalid signed JWT expiration upon login
If the backend and the bot/BDK do not have their clocks synced, the
signed JWT generated to authenticate could be considered as invalid
because the expiration is considered to be more then 5 minutes by the
backend.

We use 4 minutes instead to accomodate that.

Given we create a new token for each authentication attempt this should
be fine to use less than 5 minutes (we expect the signed JWT to be used
quickly).