Fixed characters escape in PresentationMLParser (#572)
PresentationMLParser getTextContent method was unescaping content characters before parsing it which fails when the message contains messageML tags for instance. In this commit, we changed operations order to parse the message first with the escaped characters, then unescape it in a second time

In addition to that, we now only extract content and unescape characters and we do not support anymore html entities such as &nbsp;.
According to the presentationML renderer, it seems we never supported them.
PresentationML content should always have standard escaped xml entities.