Introducing Common JWT support for Pod APIs
Goal of this implementation is to be able to start using the common JWT to call pod public APIs.
First of all the usage of common jwt can be configured by a configuration property in the application.yaml file.
when the feature is enabled, the Authorization token will be used as header for each Pod call instead of the sessionToken. This is possible thanks to a new ApiClient called BearerEnabledApiClient. When the bot is in common jwt mode enabled we will check the expiration date found in the jwt before making any call. If this date is expired then we eill trigger the refresh token and a new common jwt will be requested and saved in the AuthSession.