Improving tokens refresh
To avoid calling the login api to many times, we now first try to refresh the expired jwt by calling the idm/tokens endpoint. if that do not work we then try to refresh all the tokens as usual.
Also AuthSessionCertImpl and AuthSessionImpl have been merged since they were executing the same code