Target specific commit for symphony-api-specs
Goal of this PR is to start targetting a specific commit of the Symphony-api-spec repository and not always target the master branch, this way we'll have more controll over compatibility and what features are included in each BDK version.
Also since from 20.14 we are starting deprecation of apps authentication endpoint using certificate, we are going to generate the code from the authenticator-api-public-deprecated.yaml file so that we still support those apis.