Starting Datafeed v1 deprecation process
The datafeed 1 service will be fully replaced by the datafeed 2 service in the future. Please consider migrating over to datafeed 2. For more information on the timeline as well as on the benefits of datafeed 2, please reach out to your Technical Account Manager or to our developer documentation https://docs.developers.symphony.com/building-bots-on-symphony/datafeed
This PR includes:
- if datafeed version is not defined in the config, the BDK will use DFv2 by default
- added deprecation notice when DFv1 is used (both in the code and in the documentation)
- APIs generation is been updated following swagger deprecation of the APIs https://github.com/finos/symphony-api-spec/pull/143