Fix the event timestamp value is not accessible bug (#748)
The event timestamp is filled at the event wrapper level only, which is
the one when the event is read from datafeed. However only the raw event
is pushed to the real time listeners, that's the reason why we loose
this event timestamp.

This commit fix this issue by constructing an event proxy at runtime,
all events implements EventPayload interface, the timestamp value is
now accessible through this interface. User has to cast the raw event
to EventPayload to read the value. This big advantage is we do not
bring in any breaking changes in any stable APIs, the change is totally
transparent for the BDK developers.