Make Activity api offer asynchronous processing option (#760)
Activity api, not being asynchronous, leads to a blocking issue where a command takes long time to get processed and blocks all incoming commands. In this PR, we make the activity api support asynchronous processing. This is reached by passing a new option asynchronous to the slash annotation. In this case, an executor service is used to provide a separate thread to process the command and leaves the main thread free to receive new commands.

The executor service uses "newCachedThreadPool" to avoid creating a new thread pool for each command as it will reuse the previously finished ones. The number of threads is unbounded by using this constructor. A factory ThreadFactoyBuilder is introduced to isolate the threads creation from the core code of AbstractActivity.