Update health check strategy (#795)
Separte DataFeedLoop health check from the other components,
set the global status to out of service only if DataFeedLoop
is down. In this case, when calling healtch check endpoint,
the resposne http code is 500, otherwise if any other components
is down, set the global status to "WARNING", the http code of the response
stays as 200.

This will avoid unncessary BDK Bot restarting when other components
are not healthy.